Removed exception handler around self.data.get() from __getitem__().
User jtatum had played with this a bit and could not get an exception to raise.
To help with future debugging, removing the exception clause for now.  In the
event an exception is encountered in the future a more specific exception
handler can be put back in (vs the catch-all currently in use).