Fix error when attempting to press function keys.
Function key constants were uppercase, which triggered
BaseAXUIElement._addKeyToQueue to attempt to send the
lowercase version of the constant along with the shift
key.  This kept those keys from being found in the keymap
and resulted in an error ("ValueError: Please provide
only one character to send") when trying to send any
function key.