Add a class method to get the current frontmost app
Add a class method to get the current frontmost app
at the suggestion of a user.  This allows a user
to find what the current frontmost app is.

Current implementation is essentially a copy of
getAppRefByLocalizedName(), so some refactoring
would be in order.

Testing done:

>>> import atomac
>>> sys.path.insert(0,
... '/Users/awu/Documents/projects/pyatom/build/lib.macosx-10.6-universal-2.6/atomac')
>>> sys.path.pop(7)
'/Library/Python/2.6/site-packages/atomac-0.9.3-py2.6-macosx-10.6-universal.egg'
>>>
>>> atomac.getFrontmostApp()
<atomac.AXClasses.NativeUIElement AXApplication u'Terminal'>
>>>
>>> time.sleep(1); atomac.getFrontmostApp()
<atomac.AXClasses.NativeUIElement AXApplication u'Safari'>